src/worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "shared",
    "celery[rabbitmq]>=5.3",
    "tensorflow>=2.15",
    "faster-whisper>=1.0",
    "easyocr>=1.7",
    "open-clip-torch>=2.24",
    "sentence-transformers>=3.0",
    "boto3>=1.34",
    "opencv-python-headless>=4.9",
    "ffmpeg-python>=0.2",
    "numpy>=1.26",
    "structlog>=24.1",
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0.30",
    "pgvector>=0.3.2",
    "httpx>=0.27",
]

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["worker"]
